better auth + posts + init db script

+++ routers/influencers.py
@@ -1,15 +1,17 @@
-from fastapi import Depends, APIRouter
+from fastapi import Depends, APIRouter, HTTPException
 from sqlalchemy.orm import Session
 
 from deps import get_db, get_current_user
 from models import Influencer, Campaign
 from schemas.influencer_schema import InfluencerSchema, InfluencerCreateSchema
+from schemas.general_response_schemas import GeneralBoolResponseSchema
+
 from datetime import datetime
 
 router = APIRouter()
 
 
-@router.get("/influencers/uid/{user_id}", response_model=list[InfluencerSchema])
+@router.get("/influencers/", response_model=list[InfluencerSchema])
 def get_influencers_by_user(user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> list[Influencer]:
     res = db.query(Influencer).filter(Influencer.user_id == user.id).all()
     return res
@@ -17,18 +19,20 @@ def get_influencers_by_user(user: str = Depends(get_current_user), db: Session =
 
 @router.get("/influencers/id/{influencer_id}", response_model=InfluencerSchema)
 def get_influencer_by_id(influencer_id: int, user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> Influencer:
-    res = db.query(Influencer).filter(Influencer.id == influencer_id).first()
+    res = db.query(Influencer).filter(Influencer.user_id ==
+                                      user.id, Influencer.id == influencer_id).first()
     if not res:
-        raise Exception("influencer not found")
+        raise HTTPException(status_code=404, detail="Influencer not found")
+
     return res
 
 
 @router.get("/influencers/campaign/{campaign_id}", response_model=list[InfluencerSchema])
 def get_influencers_by_campaign(campaign_id: int, user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> list[Influencer]:
     campaign: Campaign = db.query(Campaign).filter(
-        Campaign.id == campaign_id).first()
+        Campaign.id == campaign_id, Campaign.user_id == user.id).first()
     if not campaign:
-        raise Exception("campaign not found")
+        raise HTTPException(status_code=404, detail="Campaign not found")
     influencer_ids = [influ.id for influ in campaign.influencers]
     result = db.query(Influencer).filter(
         Influencer.id.in_(influencer_ids)).all()
@@ -39,16 +43,16 @@ def get_influencers_by_campaign(campaign_id: int, user: str = Depends(get_curren
 def create_influencer(influencer_schema: InfluencerCreateSchema, user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> Influencer:
     added = datetime.now()
     influencer = Influencer(name=influencer_schema.name,
-                            note=influencer_schema.note, date_added=added, user_id=influencer_schema.user_id)
+                            note=influencer_schema.note, date_added=added, user_id=user.id)
     db.add(influencer)
     db.commit()
     db.refresh(influencer)
     return influencer
 
 
-@router.delete("/influencers/{influencer_id}")
-def remove_influencer(influencer_id: int, user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> bool:
-    db.query(Influencer).filter(Influencer.id == influencer_id).delete()
+@router.delete("/influencers/{influencer_id}", response_model=GeneralBoolResponseSchema)
+def remove_influencer(influencer_id: int, user: str = Depends(get_current_user), db: Session = Depends(get_db)) -> GeneralBoolResponseSchema:
+    db.query(Influencer).filter(Influencer.id == influencer_id, Influencer.user_id == user.id).delete()
     db.commit()
     db.commit()
-    return True
+    return GeneralBoolResponseSchema(success=True)
